+++ perllib/FixMyStreet/Cobrand/Greenwich.pm
@@ -113,9 +113,14 @@ sub open311_extra_data_include {
     ];
 
     if ($row->geocode) {
-        my $address = $row->geocode->{resourceSets}->[0]->{resources}->[0]->{address};
+        my $address;
+        if ($row->geocode->{resourceSets}) { # Bing results
+            $address = $row->geocode->{resourceSets}->[0]->{resources}->[0]->{address}->{formattedAddress};
+        } else {
+            $address = $row->geocode->{display_name}; # OSM results
+        }
         push @$open311_only, (
-            { name => 'closest_address', value => $address->{formattedAddress} }
+            { name => 'closest_address', value => $address }
         );
         $h->{closest_address} = '';
     }