+++ perllib/FixMyStreet/Geocode/Address/OSM.pm
@@ -1,0 +1,9 @@
+package FixMyStreet::Geocode::Address::OSM;
+
+use parent 'FixMyStreet::Geocode::Address';
+
+sub brand { "OpenStreetMap" }
+
+sub name_key { "display_name" }
+
+1;