+++ CHANGELOG.md
@@ -7,6 +7,8 @@
         - Stop map panning breaking after long press. #4423
     - Development improvements:
         - Extra data columns now stored as JSON, not RABX. #3216
+    - Changes:
+        - Switch to OpenStreetMap for reverse geocoding. #4444
 
 * v5.0 (10th May 2023)
     - Front end improvements: